requirements.txt
Django==2.1.5
django-model-utils==3.1.2
gunicorn==19.9.0
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.9
